requirements.txt
fastapi
python-multipart
python-dotenv
supabase
uvicorn
openai-whisper
torch
numpy
requests
pydantic
transformers
peft
whisper
